requirements.txt
Django==1.9.5
django-rest-swagger==0.3.3
djangorestframework==3.3.3
MySQL-python==1.2.5
PyYAML==3.11
readline==6.2.4.1
